refactor: outsorce logic for infinite scrolling into custom hook
